# Fix non-relative samples