Closes #3533

This adds a `fixPaths` function to `DataFile` to attempt to:

* Provide an upgrade path for projects created with 1.2.0-rc1 or 1.2.0-rc2
* Fix non-relative samples which will occur if a project containing external samples is accidentally saved using a custom lmms working directory prior to setting the preference.

Easiest way to test this out is to re-save an existing `1.2.0-rc2` project as an `.mmp` file (without the "z") and validate the full paths for built-in samples or lmms working directory samples have been corrected.