@Umcaruje 2017-05-07
Did you test out if this affects the loading speed of projects? Running this on every project load seems to be a bit much from a quick glance.