@tresf 2017-05-07
DOM parsing is slow in general so I would expect a slight uptick in load time.

A version check can be added to mitigate this, but then we would move it back to the upgrade area.  This would remove the usefulness it may offer in edge cases where the working directory is mucked up so it is a trade-off that requires a decision.