@tresf 2017-05-07
This function would eventually handle the same problem with VSTs and other plugins suffering identical edge cases.

If someone wants to benchmark it please do.