@tresf 2017-05-08
> Did you test out if this affects the loading speed of projects?

Yes.  There does not seem to be a measurable difference.  Note, tests were performed against Unfa - Spoken, which contains over 50 tracks, but very few sample tracks.

I tried again with one of my own projects containing about 12 tracks containing sample or soundfonts, and the upgrade benchmark had identical results.

Before this PR:
```
#### START UPGRADING:  "21:13:19"
#### FINISH UPGRADING:  "21:13:19"
```

After this PR:
```
#### START UPGRADING:  "21:08:34"
#### FINISH UPGRADING:  "21:08:34"
```