@tresf 2017-05-08
Ok, I noticed my benchmarks were in seconds.... Here's a proper benchmark in milliseconds... 

Without this PR:
```
#### START UPGRADING:  1494207446807
#### START UPGRADING:  1494207446830
DELTA: 4ms
SECOND TEST: 4ms
```


With this PR:
```
#### START UPGRADING:  1494207446807
#### FINISH UPGRADING:  1494207446830
DELTA: 23ms
SECOND TEST: 12ms
```

This is an 3x-6x load time, resulting in an increase in about 8-18ms; about 1/50th - 1/100th of a second on a modern (i7) machine and 50+ tracks.