@curlymorphic 2017-05-08
20ms in a non-time critical operation, compared to non-working project files, that may take a user hours to relocate the samples, if the problem is no apparent, It's a no brainer