@Umcaruje 2017-05-09
Or perhaps we could put this in a context menu somewhere, though that's gonna be hard to find if you don't know what you're looking for. 

I guess 20ms really isn't that much overload added. Did you try on some really sample intensive project that had this problem, e.g. was saved with 1.2 with broken paths, or was that the 2nd project you used? I couldn't make it out from the comment.

Real nice work on this issue though, we can release RC3 after this and other 3 issues in the project are merged :+1: