@tresf 2017-05-10
> Did you try on some really sample intensive project that had this problem

Not overly, no, but the overhead shouldn't be any different.  The logic pretty much says... 

```js
for (every single instrument) {
     if (the src tag is not relative) {
          // make it relative;
     }
}
```
DOM (XML/HTML) parsing is generally more CPU intensive than a single QString/QFileInfo operation, so the benchmarks should not increase much more with different types of instruments.  [CPU time](https://forum.qt.io/topic/70115/qdomdocument-vs-libxml2-on-performance) is going to be spent in the otherwise useless loop over elements in the DOM.

Some more information worth noting... these projects take a considerable amount of time to load.  Unfa-spoken takes about 9 seconds to load on my i7 desktop, so the overall impact (as @curlymorphic suggests) would be 20ms more, or to put plainly, 9,000ms would become 9,020ms, or 1/20th of a blink of an eye.