@Lukas-W 2017-05-10
I had a look at some of the instrument's loading code and it appears that `tryToMakeRelative` is already called on every load:

* GigPlayer: [`GigPlayer::loadSettings`](https://github.com/LMMS/lmms/blob/v1.2.0-rc2/plugins/GigPlayer/GigPlayer.cpp#L129) and [`GigPlayer::openFile`](https://github.com/LMMS/lmms/blob/v1.2.0-rc2/plugins/GigPlayer/GigPlayer.cpp#L215)
* AFP: [`audioFileProcessor::loadSettings`](https://github.com/LMMS/lmms/blob/v1.2.0-rc2/plugins/audio_file_processor/audio_file_processor.cpp#L237), [`audioFileProcessor::setAudioFile`](https://github.com/LMMS/lmms/blob/v1.2.0-rc2/plugins/audio_file_processor/audio_file_processor.cpp#L336), [`SampleBuffer::setAudioFile`](https://github.com/LMMS/lmms/blob/v1.2.0-rc2/src/core/SampleBuffer.cpp#L1187)
* SampleTCO: [`SampleTCO::loadSettings`](https://github.com/LMMS/lmms/blob/v1.2.0-rc2/src/tracks/SampleTrack.cpp#L263), [`SampleTCO::setSampleFile`](https://github.com/LMMS/lmms/blob/v1.2.0-rc2/src/tracks/SampleTrack.cpp#L151)

Similar code can be found in instruments Patman and SF2 Player.

A quick test on current `stable-1.2` confirmed this at least with sample tracks: Opening a file with broken paths will call `tryToMakeRelative` and fix the path on loading, saving it will write the correct relative paths (when using the same installation the file was created with.

So it appears this PR may not even be needed…