@tresf 2017-05-11
> So it appears this PR may not even be needed…

Confirmed.  My own unit tests passed as well.  Closing.